+++ rtos/rtos.uvprojx
@@ -391,6 +391,13 @@
         </Group>
         <Group>
           <GroupName>Thread</GroupName>
+          <Files>
+            <File>
+              <FileName>rt_task.c</FileName>
+              <FileType>1</FileType>
+              <FilePath>.\rt_task.c</FilePath>
+            </File>
+          </Files>
         </Group>
         <Group>
           <GroupName>Memory</GroupName>
@@ -807,6 +814,13 @@
         </Group>
         <Group>
           <GroupName>Thread</GroupName>
+          <Files>
+            <File>
+              <FileName>rt_task.c</FileName>
+              <FileType>1</FileType>
+              <FilePath>.\rt_task.c</FilePath>
+            </File>
+          </Files>
         </Group>
         <Group>
           <GroupName>Memory</GroupName>
